refactor : RegisterButtonContainer 에서 event 함수 주입받도록 변경
