+++ src/components/RegisterForm/RegisterButtonContainer.js
@@ -1,15 +1,10 @@
-import Restaurant from "../../domain/Restaurant";
 import { clearError } from "../../utils/clearError";
 import { clearInput } from "../../utils/clearInput";
 import { $ } from "../../utils/dom";
-import { getInfo } from "./getInfo";
 import Button from "../common/Button";
-import ErrorMessage from "../common/ErrorMessage";
 import createElement from "../../utils/createElement/createElement";
-import renderFilteredRestaurants from "../../ui/renderFilteredRestaurant.js";
-import Persistence from "../../domain/persistence/Persistence.js";
 
-const RegisterButtonContainer = (restaurantList) => {
+const RegisterButtonContainer = (restaurantList, { onClickAddButton }) => {
   const cancelButton = Button({
     text: BUTTON_TEXT.CANCEL,
     style: "button--secondary",
@@ -21,12 +16,7 @@ const RegisterButtonContainer = (restaurantList) => {
   const addButton = Button({
     text: BUTTON_TEXT.ADD,
     style: "button--primary",
-    onClick: (e) => {
-      registerRestaurant(e, restaurantList);
-      if (Persistence.loadTabInfo() === "all") {
-        renderFilteredRestaurants(restaurantList);
-      }
-    },
+    onClick: (e) => onClickAddButton(e, restaurantList),
     id: "register-button",
   });
 
@@ -50,25 +40,4 @@ const closeModal = () => {
   $("#register-modal-backdrop").classList.remove("open");
   clearInput("#register-form");
   clearError();
-};
-
-const registerRestaurant = (e, restaurantList) => {
-  e.preventDefault();
-  try {
-    const info = getInfo();
-    const restaurant = new Restaurant(info);
-
-    restaurantList.add(restaurant);
-    Persistence.saveRestaurantList(restaurantList.value);
-
-    $("#register-modal-backdrop").classList.remove("open");
-
-    clearInput("#register-form");
-  } catch (e) {
-    console.log(e.message);
-
-    const currentInputField = $(`#${e.cause}-form-item`);
-
-    currentInputField.appendChild(ErrorMessage(e.message));
-  }
 };
